+++ lib/src/types.dart
@@ -1,0 +1,6 @@
+part of 'api.dart';
+
+typedef GhostPostResponse = ({
+  PaginationInfo paginationInfo,
+  List<GhostPost> posts
+});